Parallel processing form of batch_snf

Parallel processing form of batch_snf

Usage

parallel_batch_snf(
  data_list,
  distance_metrics_list,
  clust_algs_list,
  settings_matrix,
  weights_matrix,
  similarity_matrix_dir,
  return_similarity_matrices,
  processes
)

Arguments

data_list

A nested list of input data from generate_data_list().

distance_metrics_list

An optional nested list containing which distance metric function should be used for the various feature types (continuous, discrete, ordinal, categorical, and mixed). See ?generate_distance_metrics_list for details on how to build this.

clust_algs_list

List of custom clustering algorithms to apply to the final fused network. See ?generate_clust_algs_list.

settings_matrix

matrix indicating parameters to iterate SNF through.

weights_matrix

A matrix containing feature weights to use during distance matrix calculation. See ?generate_weights_matrix for details on how to build this.

similarity_matrix_dir

If specified, this directory will be used to save all generated similarity matrices.

return_similarity_matrices

If TRUE, function will return a list where the first element is the solutions matrix and the second element is a list of similarity matrices for each row in the solutions_matrix. Default FALSE.

processes

Number of parallel processes used when executing SNF.