Plot a co-clustering heatmap across multiple settings_matrix rows

Generate a heatmap using data from multiple cocluster_data objects. See ?generate_cocluster_data for more information. The resulting heatmap shows how often two patients clustered together off of several resamplings of the data.

Usage

pooled_cocluster_heatmap(cocluster_list)

Arguments

cocluster_list

A list of cocluster_data objects (from ?generate_cocluster_data) to pool together for the heatmap.