Heatmap of p-values

Heatmap of p-values

Usage

pval_heatmap(
  pvals,
  order = NULL,
  cluster_columns = TRUE,
  cluster_rows = FALSE,
  show_row_names = FALSE,
  show_column_names = TRUE,
  min_colour = "red2",
  mid_colour = "lightyellow",
  max_colour = "slateblue4",
  legend_breaks = c(0, 0.5, 1),
  col = circlize::colorRamp2(legend_breaks, c(min_colour, mid_colour, max_colour)),
  heatmap_legend_param = list(color_bar = "continuous", title = "p-value", at = c(0, 0.5,
    1)),
  rect_gp = grid::gpar(col = "black"),
  column_split_vector = NULL,
  row_split_vector = NULL,
  column_split = NULL,
  row_split = NULL,
  ...
)

Arguments

pvals

A matrix of p-values.

order

Numeric vector containing row order of the heatmap.

cluster_columns

Whether columns should be sorted by hierarchical clustering.

cluster_rows

Whether rows should be sorted by hierarchical clustering.

show_row_names

Whether row names should be shown.

show_column_names

Whether column names should be shown.

min_colour

Colour used for the lowest value in the heatmap.

mid_colour

Colour used for the middle value in the heatmap.

max_colour

Colour used for the highest value in the heatmap.

legend_breaks

Numeric vector of breaks for the legend.

col

Colour function for ComplexHeatmap::Heatmap()

heatmap_legend_param

Legend function for ComplexHeatmap::Heatmap()

rect_gp

Cell border function for ComplexHeatmap::Heatmap()

column_split_vector

Vector of indices to split columns by.

row_split_vector

Vector of indices to split rows by.

column_split

Standard parameter of ComplexHeatmap::Heatmap.

row_split

Standard parameter of ComplexHeatmap::Heatmap.

...

Additional parameters passed to ComplexHeatmap::Heatmap.