Convert a data list to a similarity matrix through a variety of SNF schemes

Convert a data list to a similarity matrix through a variety of SNF schemes

Usage

snf_step(
  data_list,
  scheme,
  k = 20,
  alpha = 0.5,
  t = 20,
  cont_dist_fn,
  disc_dist_fn,
  ord_dist_fn,
  cat_dist_fn,
  mix_dist_fn,
  weights_row
)

Arguments

data_list

A nested list of input data from generate_data_list().

scheme

Which SNF system to use to achieve the final fused network

k

k hyperparameter

alpha

alpha/eta/sigma hyperparameter

t

SNF number of iterations hyperparameter

cont_dist_fn

distance metric function for continuous data

disc_dist_fn

distance metric function for discrete data

ord_dist_fn

distance metric function for ordinal data

cat_dist_fn

distance metric function for categorical data

mix_dist_fn

distance metric function for mixed data

weights_row

dataframe row containing feature weights

Value

fused_network The final fused network for clustering