Create subsamples of a data_list

Given a data list, return a list of smaller data lists that are generated through random sampling (without replacement).

Usage

subsample_data_list(
  data_list,
  n_subsamples,
  subsample_fraction = NULL,
  n_subjects = NULL
)

Arguments

data_list

A nested list of input data from generate_data_list().

n_subsamples

Number of subsamples to create.

subsample_fraction

Percentage of patients to include per subsample.

n_subjects

Number of patients to include per subsample.

Value

A "list" class object containing n_subsamples number of data lists. Each of those data lists contains a random subsample_fraction fraction of the observations of the provided data list.