Run SNF clustering pipeline on a list of subsampled data lists.

Run SNF clustering pipeline on a list of subsampled data lists.

Usage

batch_snf_subsamples(
  data_list_subsamples,
  settings_matrix,
  processes = 1,
  return_similarity_matrices = FALSE,
  clust_algs_list = NULL,
  suppress_clustering = FALSE,
  distance_metrics_list = NULL,
  weights_matrix = NULL,
  automatic_standard_normalize = FALSE,
  return_solutions_matrices = FALSE
)

Arguments

data_list_subsamples

A list of subsampled data lists. This object is generated by the function batch_snf_subsamples().

settings_matrix

A settings matrix defining the parameters of the SNF pipelines to be applied to the subsampled data lists.

processes

See ?batch_snf.

return_similarity_matrices

See ?batch_snf.

clust_algs_list

See ?batch_snf.

suppress_clustering

See ?batch_snf.

distance_metrics_list

See ?batch_snf.

weights_matrix

See ?batch_snf.

automatic_standard_normalize

See ?batch_snf.

return_solutions_matrices

If TRUE, includes the solutions matrices corresponding to each subsample in the output.