Manhattan plot of feature-meta cluster associaiton p-values

Given a dataframe of representative meta cluster solutions (see get_representative_solutions(), returns a Manhattan plot for showing feature separation across all features in provided data/target_lists.

Usage

mc_manhattan_plot(
  extended_solutions_matrix,
  data_list = NULL,
  target_list = NULL,
  variable_order = NULL,
  neg_log_pval_thresh = 5,
  threshold = NULL,
  point_size = 5,
  text_size = 20,
  plot_title = NULL,
  xints = NULL,
  hide_x_labels = FALSE,
  domain_colours = NULL
)

Arguments

extended_solutions_matrix

A solutions_matrix that contains "_pval" columns containing the values to be plotted. This object is the output of extend_solutions().

data_list

List of dataframes containing data information.

target_list

List of dataframes containing target information.

variable_order

Order of features to be displayed in the plot.

neg_log_pval_thresh

Threshold for negative log p-values.

threshold

p-value threshold to plot horizontal dashed line at.

point_size

Size of points in the plot.

text_size

Size of text in the plot.

plot_title

Title of the plot.

xints

Either "outcomes" or a vector of numeric values to plot vertical lines at.

hide_x_labels

If TRUE, hides x-axis labels.

domain_colours

Named vector of colours for domains.